Initialize a Standard Library
Exn: As Preface uses (for [Try] and [Validate]) biased versions with
exceptions to describe an error, all exceptions propagated in the
project are centralised in a module [Exn] which provides equalities and
pretty-printing functions.

Pp: Some small helpers for formatter/pretty printer

Provider: Describes a uniform way of querying a key value environment
and applying validation.