Dummy endpoint with Dream and rework Lib_db
This commit allows Dream to be used to render the Dummy page and the
Registration page (only) and changes the logic for handling SQL queries.
This time `use` is only present for the binary and for integration
testing. The current approach of Lib_db was to "wear out" the consumption of the
database on each request, which is problematic in several ways:

1. This is a bit of a violation of the Demeter principle. Indeed, the
SQL consumer must explicitly manipulate the pool and have knowledge of
the Handler.

2. This is not compatible with Dream which logically only handles one
connection per request, the pool logic is abstracted via middleware.