Reimplementing registration controller
This commit reimplements the view/controller pair currently present on
`main` using Dream. The handler logic is not significantly different
from the previous version. The main (trivial) difference is the page
forwarding logic.

Previously, the page was reserved by passing it the notification to be
called. The problem with this approach is that it leaves no real control
over HTTP return codes (out of scope for this PR) and does not empty the
POST parameter buffer. As a result, despite the registration of a user,
the refresh of the page (Dummy) after saving would restart the HTTP
request. In the context of user registration, this is not
dramatic (although it is not very UX) because the user cannot be
registered twice. But in the case of a message (for example) it could be
annoying.

Now, rather than "re-rendering" the page, the server performs a
redirect. And the notification logic is handled via Flash messages.
Specifically, depending on whether the action is successful or not, the
server serializes a message (only for the "continue" request) and
performs a redirect. For each page, the message must be fetched (to
avoid `Lib_ui` depending on Dream), its reception will deserialize it,
hence the little modification in the original code.

Currently, serialization is done in JSON, but it would be possible, in
the near future, to switch to S-Expression or to use Data_encoding. But
in a policy of reducing Diff, this seemed to me to be the most relevant
choice.