A new Form validation library
The validation of the forms was based on the `Assoc` module which, being
thought to act on recursive structures (like trees, to treat JSON,
S-Expressions or Yaml), a pass to an associative list was mandatory.
This generated a lot of Boilerplate, and made the code of the models a
bit, in my humble opinion, obese.

Lib_form is a less general form (but one that fits more with Dream's
HTTP parameter handling) for building applicative validation pipelines
for Muhokama. Many opinionated choices have been implemented to allow
easy composition between a validation and a validation function,
respecting this scheme: `field_operation &> (additional validators)`.

It would be possible to greatly improve the library but this would imply
changes in Preface.

Co-authored-by: xvw <xavier.vdw@gmail.com>
Co-authored-by: xhtmlboy <xhtmlboi@gmail.com>